+++ resources/views/article/show.blade.php
@@ -10,12 +10,16 @@
             <h5 class="card-title">{{ $article->name }}</h5>
             <p class="card-text">{{ $article->desc }}</p>
             <div class="d-flex justify-content-center">
+                @can('edit', $article->id)
                 <a href="/article/{{$article->id}}/edit" class="btn btn-primary me-2">Edit article</a>
+                @endcan
+                @can('delete', $article->id)
                 <form action="/article/{{$article->id}}" method="POST">
                     @method("DELETE")
                     @csrf
                     <button type="submit" class="btn btn-danger">Delete article</button>
                 </form>
+                @endcan
             </div>
         </div>
 
@@ -68,8 +72,10 @@
                             <p class="card-text">{{ $comment->desc }}</p>
                         </div>
                         <div class="d-flex justify-content-end card-body pt-0">
-                            <a href="/comment/{{$comment->id}}/edit" class="btn btn-primary me-2">Edit comment</a>
-                            <a href="/comment/{{$comment->id}}/delete" class="btn btn-danger me-2">Delete comment</a>
+                            @can('update_comment', $comment)
+                                <a href="/comment/{{$comment->id}}/edit" class="btn btn-primary me-2">Edit comment</a>
+                                <a href="/comment/{{$comment->id}}/delete" class="btn btn-danger me-2">Delete comment</a>
+                            @endcan
                         </div>
                     </div>
                 @empty
